remove tree module
