+++ src/lib.rs
@@ -44,7 +44,6 @@ pub mod status;
 pub mod tab;
 pub mod term_manager;
 pub mod trash;
-pub mod tree;
 pub mod trees;
 pub mod users;
 pub mod utils;
